requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-markdownx==3.0.1
django-sass==1.0.1
idna==2.10
libsass==0.20.1
Markdown==3.3.4
martor==1.6.2
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
